[pipeline/rpc] implement distributed optimizer | test with assert_close
